Update encoders files
